refactor graphQL types
