+++ graphQL/resolvers/index.js
@@ -1,0 +1,8 @@
+const postsResolvers = require('./posts')
+const usersResolvers = require('./users')
+
+module.exports = {
+    Query: {
+        ...postsResolvers.Query
+    }
+}
