+++ index.js
@@ -1,37 +1,9 @@
 const {ApolloServer} = require('apollo-server')
-const gql = require('graphql-tag')
 const mongoose = require('mongoose')
-
+const typeDefs = require('./graphQL/typeDefs')
 
 const {MONGODB} = require('./config.js')
-const Post = require('./models/Post')
-
-const typeDefs = gql`
-type Post{
-    id: ID!
-    body: String!
-    createdAt: String!
-    username: String!
-}
-type Query {
-    # sayHi: String!, 
-    getPosts: [Post]
-}
-`
-
-const resolvers = {
-    Query: {
-        // sayHi: () => 'Hello World'
-        async getPosts(){
-            try {
-                const posts = await Post.find()
-                return posts
-            } catch (err) {
-                throw new Error(err)
-            }
-        }
-    }
-}
+const resolvers = require('./graphQL/resolvers')
 
 const server = new ApolloServer({
     typeDefs,
